Revert "feat(calculator): 그래프 x축 수정 및 과제 20개 이상 추가시 에러모달 추가"
This reverts commit e52447f2a2d63dd2c5e401f7452a470248551eec.
