--- app/src/@core/atoms/calculatorDialogAtom.ts
@@ -1,6 +1,0 @@
-import { atom } from 'jotai';
-
-export const calculatorDialogAtom = atom({
-  isOpen: false,
-  description: '',
-});
