+++ app/src/@core/atoms/isSubjectDuplicateAtom.ts
@@ -1,0 +1,3 @@
+import { atom } from 'jotai';
+
+export const isSubjectDuplicateAtom = atom(false);
