+++ app/src/@core/providers/ModalProvider.tsx
@@ -1,21 +1,21 @@
 import { isSpotlightOpenAtom } from '@core/atoms/isSpotlightOpenAtom';
 import { reLoginDialogInfoAtom } from '@core/atoms/reLoginDialogInfoAtom';
-import { calculatorDialogAtom } from '@core/atoms/calculatorDialogAtom';
+import { isSubjectDuplicateAtom } from '@core/atoms/isSubjectDuplicateAtom';
 import { ReLoginDialog } from '@core/components/Modal/ReLoginDialog';
-import { CalculatorDialog } from '@core/components/Modal/CalculatorDialog';
+import { SubjectDuplicateDialog } from '@core/components/Modal/SubjectDuplicateDialog';
 import { Spotlight } from '@core/components/Spotlight';
 import { PropsWithReactElementChildren } from '@shared/types/PropsWithChildren';
 import { useAtomValue } from 'jotai';
 
 const ModalProvider = ({ children }: PropsWithReactElementChildren) => {
   const { isOpen: isReLoginDialogOpen } = useAtomValue(reLoginDialogInfoAtom);
-  const { isOpen: isCalculatorDialogOpen } = useAtomValue(calculatorDialogAtom);
   const isSpotlightOpen = useAtomValue(isSpotlightOpenAtom);
+  const isSubjectDuplicate = useAtomValue(isSubjectDuplicateAtom);
   return (
     <>
       {isReLoginDialogOpen ? <ReLoginDialog /> : null}
       {isSpotlightOpen ? <Spotlight /> : null}
-      {isCalculatorDialogOpen ? <CalculatorDialog /> : null}
+      {isSubjectDuplicate ? <SubjectDuplicateDialog /> : null}
       {children}
     </>
   );
