+++ app/src/Calculator/CalculatorInputMobile.tsx
@@ -8,12 +8,11 @@ import {
   H3MediumText,
 } from '@shared/ui-kit';
 import styled from '@emotion/styled';
-import { useAtom, useSetAtom } from 'jotai';
+import { useAtom } from 'jotai';
 import { subjectListAtom } from '@/Calculator/atoms/subjectListAtom';
 import { ProjectSpotlight } from '@/Calculator/ProjectSpotlight';
 import { calculatorPropsAtom } from './atoms/calculatorPropsAtom';
 import { calculateSubjectList } from '@/Calculator/utils/calculateSubjectList';
-import { calculatorDialogAtom } from '@core/atoms/calculatorDialogAtom';
 import { OrderItemButton } from '@/Calculator/OrderItemButton';
 import { TableRowList, Subject } from '@/Calculator/types/orderItemButton';
 import { Button } from '@shared/ui-kit';
@@ -21,7 +20,6 @@ import { Button } from '@shared/ui-kit';
 const CalculatorInputMobile = () => {
   const [subjectList, setSubjectList] = useAtom(subjectListAtom);
   const [calculatorProps] = useAtom(calculatorPropsAtom);
-  const setCalculatorDialogAtom = useSetAtom(calculatorDialogAtom);
   const currentLevel = calculatorProps.currentLevel;
 
   const onListChange = (subjectList: TableRowList[]) => {
@@ -33,13 +31,7 @@ const CalculatorInputMobile = () => {
   };
 
   const onAddClick = () => {
-    if (subjectList.length >= 20) {
-      setCalculatorDialogAtom({
-        isOpen: true,
-        description: '과제는 최대 20개까지 추가 가능합니다.',
-      });
-      return;
-    }
+    if (subjectList.length >= 20) return; //ToDo: 20개 이상 추가 불가능 안내 추가
     const calculatedSubjectList = calculateSubjectList({
       subjectList: subjectList,
       currentLevel: currentLevel,
