--- app/src/Calculator/constants/level.ts
@@ -1,2 +1,0 @@
-export const MAX_XAXIS_COUNT = 8;
-export const MAX_XVALUE_LENGTH = 20;
